mononoke: fix panic in derived data
Summary:
Derived data has a concept of a mapping, which is a persistent key-value store,
where Key is a bonsai changeset id, and Value depends on particular derived
data implementation (quite often it's a hash of the content). The mapping
itself is usually stored in a database or a blobstore.

Among other things derived data mapping is used to find out whether derived
data for this particular bonsai commit was derived or not. One important caveat
though is that **it's not guaranteed that ancestor changeset has a mapping
entry if descendant changeset has a mapping entry**. This can happen because
mapping entries for a parent and a child are inserted in parallel, and this was done
for performance reasons - inserting mapping entries sequentially might become a bottleneck for
deriving for long stacks. Note that this optimization doesn't affect
correctness, however it might cause that some commits might derived a few
times.

Derived data framework topologically sorts commits to derive ancestors before
descendantas. Now consider the following case:

```
A <- no mapping
|
B <- mapping exists
|
C <- no mapping
```

There are a few key contributors to this bug:
1) The code that looks for commits to derive looks for a set of commits that is bounded by derived commits. Here, that's A.
2) Topological sort will return a topologically set of commits and their parents. That means A and B.

Those two facts put together means that derived data will try to derive commit B and then looks at the mapping for commit C which doesn't exist and hence it will panic.

This diff fixes it by making sure we don't try to derive a commit for commits
like "B".

Reviewed By: krallin

Differential Revision: D19298659

fbshipit-source-id: f3559074611370ebb5ddf2c728b3620fa30c17f8