mononoke: add the ability to set the number of theads in tokio Runtime from CLI
Summary:
The default behavior of `tokio::runtime::Runtime` is to use the number of CPUs
as per `num_cpus::get().max(1)` (see [1], [2]). This is a reasonable default on
physical systems with unlimited resource access, but in containerized
environment Mononoke may be limited to a smaller number of CPUs than totally
available on a system. Let's add the ability to override the size of the thread
pool on the command line.

[1] https://github.com/tokio-rs/tokio/blob/tokio-udp-0.1.2/tokio-threadpool/src/builder.rs#L93
[2] https://github.com/tokio-rs/tokio/issues/400

Reviewed By: farnz

Differential Revision: D19309494

fbshipit-source-id: 1fa4054b940bf225fac5f1a518f5cbb50d882fda