mononoke: walker: move scuba table up to common setup
Summary:
move scuba table definition up to common setup as it makes sense to be able to run various walks with scrub blobstore settings, in which case we want scrub failures logged and common value for walk_type.

Also make the subcommand walk_types constants to reduce risk of typos.

Reviewed By: StanislavGlebik

Differential Revision: D19226120

fbshipit-source-id: 64553d2aa2d55eb413360b4e301b7765a0f5dc2a