mononoke: filenodes: use smaller chunks on Sqlite
Summary:
Sqlite has limits on the number of parameters in a query. Let's account for
this here by using smaller chunks. A slightly better approach might be to do
group those by shard first, but considering what we have has worked fine for
now, I've opted for minimizing the scope of the change here.

Reviewed By: ahornby

Differential Revision: D19327329

fbshipit-source-id: ccfe0e0953f5cf8437b3ce2692e2bb3127c9b773