mononoke: make find_bookmark_diff() easier to use
Summary:
Previously find_bookmark_diff() only worked with CommitSyncer structure that
maps from large repo to a small repo. That was very confusing, since it wasn't
clear from the function signature.

This diff changes it, and now find_bookmark_diff() works in both direction i.e.
it can verify small -> large direction and large -> small direction. The main
caveat is that we need to provide reverse_bookmark_mapping for this function to
work correctly (see code comments for more details). While it's still not
great, I'd argue it's a step in the right direction compared to what we had
before.

While at it, I also move 2 out of 3 tests for find_bookmark_diff() into validation.rs
file, so that the tests reside closer to the code. There's still one test left
in admin/crossrepo.rs because it uses functionality that's doesn't exist in
validation.rs.

Reviewed By: ikostia

Differential Revision: D19314399

fbshipit-source-id: 4a3124e3645d0f16762a83f7d4b9e311490b21b4