mononoke/lfs_server: fixup test-lfs-server-live-config.t
Summary:
bruy, while working on a bootcamp in this area noted that this test wasn't
doing anything useful. The test updates the live config, then ... checks that
it has *not* updated (which is the exact opposite of what we want).

This is happening because we ignore invalid configs when reloading them
(because that's better than crashing the service), and the config we put in
there in the test is indeed invalid (because it's missing a trailing comma).

We accidentally broke this in D18452653. I didn't noticed while reviewing, so
this patch 1) fixes the bug and 2) makes it harder to not notice if we break it
again.

Reviewed By: StanislavGlebik

Differential Revision: D19348023

fbshipit-source-id: 732f3d77057e8bb80e54e0473fb9bbc9b986e879