scs_server: add scuba tracing for request start and end
Summary:
Extract creation of the `CoreContext` to the macro definition of each
thrift method, and then use this `CoreContext` to log to scuba at the
start and end of each request.

In order to log whether the request failed because of a request error
or because of an internal error, all method implementations need
to return `ServiceError`.  This works to our advantage, as it means
we can remove the `Into<thrift::Error>` impls for `MononokeError`
in the `mononoke_api` crate.

Reviewed By: krallin

Differential Revision: D19345597

fbshipit-source-id: 559be4b73b52d501803702d0dc180f7cf3136492