sshpeer: read stderr in a background thread
Summary:
On Windows, select cannot be used for pipes, and thus the doublepipe class
would end up busy spinning trying to read from it. Since stderr is almost
always empty let's move it to a background thread so the main code path
doesn't have to deal with it.

This proves to be a significant win for `hg pull`. For an empty manifest
cache, `hg pull` used to take ~5 minutes, it's now around 1 minute, with
Mononoke taking 40 seconds out of that.

Reviewed By: DurhamG

Differential Revision: D19352144

fbshipit-source-id: 77112c1c18d7dab15f6946a645f73e89908802f2