mononoke: extract common functionality to find unsynced ancestors
Summary:
backsyncer uses it to find what commits still need syncing.
I'm planning to use the same funcionality in the next diff, so let's extract it.

I modified this function a bit to return only unsynced ancestors (previously it was returning unsynced ancestors and their immediate parents because of sort_topological function semantic). Thanks to that we can now remove a check about whether the commit is synced or not from backsyncer.

Reviewed By: krallin

Differential Revision: D19390626

fbshipit-source-id: d7c7783f7e063b30698f599aa26b41ec898d91c6