tracing: remove FacebookInit from TraceContext
Summary:
The FacebookInit was introduced in TraceContext here: D18660027 in order to be able to upload traces to manifold. The problem here is that this introduces the concept of FacebookInit to the trace itself rather than only to the uploading the trace.
Considering open-sourcing of this code the oss version of TraceContext would have to be aware of FacebookInit even though it won't be uploading it to manifold, so the FacebookInit wouldn't be used.

This diff fixes this design and pushes the knowledge of FacebookInit out of TraceContext to the callsites that wish to upload it to manifold.

Reviewed By: farnz

Differential Revision: D19428218

fbshipit-source-id: 19b6115f67c0de35fbf4b15c9a47fa2510afd24c