mononoke: walker: add option of error-as-data
Summary:
Add an option to allow errors loading a particular node type to be treated as data so the walk can continue (although maybe not discover all items as a path may be cut off by the error) .

This is designed to allow a walk to discover as many errors as possible rather than failing at the first one.

The new options should only be run in --read-only-storage mode to remove risk of persisting error contaminated state.  The new test checks for this.

Reviewed By: farnz

Differential Revision: D19375929

fbshipit-source-id: 438bc29f97ce90c31610ca11e748bfc18d5c1c2e