mononoke: introduce new mononoke_x_repo_sync_job
Summary:
This diff introduces new_mononoke_x_repo_sync_job. The main difference between this sync job and old sync job is that new sync job is capable of syncing all bookmarks, while old sync job was syncing only a single bookmark. It does so by following bookmark update log of the source repo.

Probably the most surprising thing is that there can be two ways to sync a commit - we can either just rewrite it and copy into the target repo, or rewrite and pushrebase on top of the corresponding bookmark in target repo. The latter is necessary for bookmarks that are common between a few small repos (usually this bookmark is "master") - a few sync jobs might update the same bookmark at the same time, hence there might be a race.

Note: while it was supposed to be an update of the old sync job, the differences were too big, and I decided to instead introduce a new crate. In the later diff I'm going to remove the old sync job and rename new_mononoke_x_repo_sync_job into mononoke_x_repo_sync_job.

Note: this job doesn't support merges completely. In the next diff I'll lift this restriction a bit to allow the simplest kind of merges to be synced (i.e. merges that introduce new repo).

Reviewed By: ikostia

Differential Revision: D19373372

fbshipit-source-id: d4ec9d0ae6b056b3cb91260ca8b7a43336c9438f