mononoke: handle merges in new x_repo_sync_job, step 2
Summary:
This is a follow up from the previous diff - it allows pushing pre-merge
commits first, and then pushing the merge commit to master separately.

The reason it wasn't possible before is because it wasn't possible to sync
merge commits to a non-common pushrebase bookmark. I.e.

```
O <- master
| \
O  O  <- new repo we are trying to merge in
   |
   O <- this new repo contains merges itself!
  / \
```

This diff allows doing syncing merges for non-common pushrebase bookmarks,
but only for the simplest cases i.e. when there's no intersection between
ancestors of common pushrebase bookmarks and ancestor of the merge we are
trying to sync.
To do that I added an unsafe function naive_sync_commit, which does a very
simple rewrite and sync of a commit from source repo to target without doing
any additional checks. This is an unconventional use of unsafe, however I think
it make sense - using this function directly can lead to repo corruption, so
it's up to the caller to make sure it doesn't happen.

Reviewed By: ikostia

Differential Revision: D19408670

fbshipit-source-id: 1b2dc68110f0a00f3b63ab183f02a993c80e7d73