multiplexedblob: don't sample puts
Summary:
We have few puts (about 0.5% of our ops are puts), so we can afford to log them
more aggressively. Considering our current sampling ration this will
effectively double the amount of we're logging — a steep increase, but a
manageable one.

As part of this diff, I've refactored how we sample things into a `Sampling`
struct that is a little nicer to use than the current approach of using the
sampling constant here and there.

Reviewed By: farnz

Differential Revision: D19430387

fbshipit-source-id: 1163bf17c2bede55e28b20c132d8f5ae19f2a360