mononoke: make crossrepo verify-bookmarks easier to use
Summary:
Previously if a commit from a large didn't have an entry in the mapping table
for a small repo, then verification would just fail and it doesn't check
positions of any other bookmarks.

However this situation is absolutely expected for bookmarks that are shared
between small repos e.g. for master. If master points to a commit from small repo A,
then it might not have an entry in the mapping table for small repo B, and so `crossrepo
verify-bookmarks` fails.

This diff updates it so that it prints the warning, but verifies positions of
all other bookmarks.

Reviewed By: farnz

Differential Revision: D19449377

fbshipit-source-id: 331605fe9450bcbc5864645dd6eb23b51e7ffa86