mononoke: support octopus merges in hg derivation through stepparents
Summary:
This updates hg derivation to minimally handle octopus merges. The underlying
goal is to allow converting a Git repository to a Mercurial repository by going
through Bonsais (and to produce something meaningful in doing so).

Since hg doesn't currently have a notion of parents beyond p1 and p2, this
support is (purposefully) very minimal. What we seek to guarantee is that the
files in the hg manifest are logically identical to what is represented  by the
Bonsai changeset. However, we do not preserve parents in filenodes referencing
stepparents (it might be confusing for hg to encounter files).

Implementation-wise, this leaves most of the existing code unchanged: the only
place where we actually to handle > 2 parents is when deriving the Manifest.
It does however remove assertions that assume 2 parents.

To make this work, I added a `Traced` wrapper in the `manifest` crate that lets
us find which parent manifest a given child manifest or filenode came from.
This lets us filter out ancestors for parents that aren't p1 or p2.

Finally, for now, I've recorded stepparents in commit extras (we just
concatenate their hexified IDs together). This seems consistent with things
like rebase_source that I've seen here or there, but let me know if not.

Reviewed By: StanislavGlebik

Differential Revision: D19348533

fbshipit-source-id: a40cb81483b0741f09c026055c719c0c1ed8e809