mononoke/bonsai_globalrev_mapping: remove some methods we're not going to use
Summary:
My current plan for globalrevs is to assign them as part of the pushrebase
step, which ensures we put them in the Bonsais we generate (latency-wise, this
should be OK since we're already doing a roundtrip to the DB for bookmarks
anyway).

However, the current methods in our Bonsai Globalrev mapping aren't very
amenable to this:

- `add_many` ignores duplicates, so if we used that, we would end up not
  noticing duplicate globalrevs, which is problematic
- `add` inserts globalrevs one at a time, which isn't desirable. Also, it has a
  lot of ad-hoc logic to compensate for the fact that it does a `INSERT IGNORE`
  when what we really need is a regular "`INSERT` and bail on duplicates".

So, let's:

- Remove `add()` altogether. It's not being used right now anyway.
- Clarify the purpose of `add_many` by renaming it to `bulk_import`, which is
  for contexts where we expect Globalrevs to be correct (i.e. backfill &
      blobimport).

Also, while in there, removing some un-necessary Vec allocations.

Reviewed By: StanislavGlebik

Differential Revision: D19496573

fbshipit-source-id: 7d3ba25be04c71f3d54e36dd54b565dfad0d6ebc