mononoke/pushrebase: port tests to async / await
Summary:
This updates the pushrebase tests to async / await as well. Since there are a
bunch of shared helpers here, I've done this in just one go.

One little behavior change is that PushrebaseError now derives Error so we can
use `?` on it in a function that returns `Error`. Functionally this doesn't
change anything.

Reviewed By: ikostia

Differential Revision: D19496349

fbshipit-source-id: 4c5021891ec5c1cab972eeb85f43cf60a87437b1