mononoke: fix get_movers and get_bookmark_renamers
Summary:
It is incorrect to read the sync direction from the config, as config only
describes the direction of sync, while movers/renamers can be used for
verification for instance. As an example, the `CommitSyncRepos` struct is
created with direction specified by the constructor (hence `source_repo` and
`target_repo` args), which does not have to match the configured direction, yet
it called `get_movers` and `get_bookmark_renamers`, which would read direction
from config.

This diff makes `direction` an explicit argument in both functions.

Reviewed By: StanislavGlebik

Differential Revision: D19518278

fbshipit-source-id: f2c01139b785e91c5b81bc6cb545aff1c11f9393