mononoke: introduce generate_filenodes method
Summary:
At the moment filenodes [1] are generated as soon as hg changeset is generated
(see get_manifest_from_bonsai() from bonsai function). This is not great for a
few reasons:

1) It makes code more complicated - we have to pass IncomepleteFilenodes
structure all the way down to derive_hg_manifest() function, we have to make
pushrebase aware of the filenodes and we have to pass "draft" flag to
CreateChangeset.

2) It's actually incorrect - we don't want to generate filenodes for draft
commits at all [2]. It works right now because we don't generate hg changesets
for commit cloud commits, but it's just a matter of time when we start doing
it [3].

The goal of this stack is to separate filenode generation from hg
changeset generation, and move filenode generation onto derived data. Filenodes
will be generated during generating getbundle or pusrebase responses.

This diff implements the core functionality - it generates a list of filenodes
for a bonsai changeset. It does so by finding all changed files/manifests for a
commits, and then fetching respective envelopes to get p1/p2/copy info.

Note that it's possible to avoid fetching file envelopes, however then we'd
have to replicate a complicated logic of finding correct filenode parents (see
https://fburl.com/25l6h1nm), and I don't think perf benefit outweighs
introduced code complexity.

Reviewed By: farnz

Differential Revision: D19501515

fbshipit-source-id: 8bf9b9fcbf832d99443a146257cd93cdf7e34130