mononoke: refactor create_getbundle_response
Summary:
We are going to change this function so that it can derive filenodes before
returning commits to the user.

This function does refactoring to make it easier. In particular, it changes
create_getbundle_response to be a future, it now always requires phases_hint,
however whether phases part is returned is still controlled by a flag [1]

This change shouldn't introduce any changes in behaviour.

[1] I wouldn't be surprised if it's possible to remove phases part completely,
but it's worth doing in a separate diff.

Reviewed By: krallin

Differential Revision: D19515803

fbshipit-source-id: 6c77686c1ba9ae166bfeeae36379957e1d0627cf