mononoke: list_file_history from fastlog
Summary:
I introduced new `list_file_history` function as an ops of the Fastlog. The function returns a stream of the history commit from the given unode entry.
Th logic is pretty much the same as in APIServer's `get_file_history`, however refactored and written that way to return a stream.
Having a stream of commits is useful because it can be filtered later in SCS by `log` (using before/after timestamps, for example).

I decided not to change APIServer to use `list_file_history` to not break anything.

Reviewed By: StanislavGlebik, farnz

Differential Revision: D19161387

fbshipit-source-id: 1c5ce7a92b3a38314b6e06ef351b2095895eb184