mononoke: add segmented_changelog crate
Summary:
Segmented changelog is a new way to represent the commit graph that
can compress and quickly iterate over section of commits that have
only one parent. [1]

There is a good amount of code already for this structure in the client.
The `dag` crate in that project has a lot of the functionality already
implemented. [2] Long term we ideally have one library shared between the
client and the server but right now the runtime environments are different
enough that it is easier to reimplement the required parts for the server.
The implementation requirements also seem different enough as we look at what
the server requires.

With segmented changelogs we will want new protocols for talking between server
and client.

This change is add a seed crate with a simple implementation for idmap, the
first component of the segmented changelog. It is just relabling commits using
continuous increasing integers.
The short term goal will be to add some simple implementations that can be used
to define a valid interface. From there we can look in parallel at protocols on
one side and proper server implementation on the other.

[1] https://github.com/facebookexperimental/eden/blob/master/eden/scm/slides/201904-segmented-changelog/segmented-changelog.pdf
[2] https://github.com/facebookexperimental/eden/tree/master/eden/scm/lib/dag

Reviewed By: krallin

Differential Revision: D19525583

fbshipit-source-id: 7e27fb4a30c901b30a3bbaccddeb91060721f407